[Feature #161213990] add endpoint to post a sale record
